+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
   - Reimplement ~15% of the functions in StarCraft 1.16.1
   - Use iscript animations from a source file instead of a binary
+  - Fix crashes caused by long strings (e.g. scenario title/description)
   - Add a new trigger action: "Execute Lua"
     * Implement a `Scenario` type
     * Implement a `Player` type